chore: refactoring codes

+++ src/nz_companies_office/models/split.py
@@ -31,9 +31,9 @@ def split_share_edges(
     val_idx = perm[n_train : n_train + n_val]
     test_idx = perm[n_train + n_val :]
 
-    train_pos = edge_index[:, train_idx]
-    val_pos = edge_index[:, val_idx]
-    test_pos = edge_index[:, test_idx]
+    train_pos = torch.LongTensor(edge_index[:, train_idx])
+    val_pos = torch.LongTensor(edge_index[:, val_idx])
+    test_pos = torch.LongTensor(edge_index[:, test_idx])
 
     return train_pos, val_pos, test_pos
 
@@ -61,14 +61,18 @@ def sample_negative_edges(
         Tuple of (val_neg, test_neg) edge indices.
 
     """
-    val_neg = negative_sampling(
-        train_pos,
-        num_nodes=(n_shareholder, n_company),
-        num_neg_samples=val_pos.shape[1],
+    val_neg = torch.LongTensor(
+        negative_sampling(
+            train_pos,
+            num_nodes=(n_shareholder, n_company),
+            num_neg_samples=val_pos.shape[1],
+        ),
     )
-    test_neg = negative_sampling(
-        torch.cat([train_pos, val_pos], dim=1),
-        num_nodes=(n_shareholder, n_company),
-        num_neg_samples=test_pos.shape[1],
+    test_neg = torch.LongTensor(
+        negative_sampling(
+            torch.cat([train_pos, val_pos], dim=1),
+            num_nodes=(n_shareholder, n_company),
+            num_neg_samples=test_pos.shape[1],
+        ),
     )
     return val_neg, test_neg
